ms_mail/requirements.txt
pydantic>=2.7.1,<2.8.0
fastapi>=0.110.3,<0.111.0
uvicorn>=0.29.0,<0.30.0
python-jose>=3.3.0,<3.4.0
cryptography>=42.0.5,<42.1.0
httpx>=0.27.0,<0.28.0
pytest>=8.2.0,<8.3.0
pytest-mock>=3.14.0,<3.15.0
pytest-cov>=5.0.0,<5.1.0
pytest-httpx>=0.30.0,<0.31.0
pytest-asyncio>=0.23.6,<0.24.0
aiofiles>=23.2.1,<23.3.0
aiosmtplib>=3.0.1,<3.1.0